refactor(EmployeeDocuments): move into Documents/onboarding/ subdirectory
Relocates EmployeeDocuments into the Employee/Documents/ directory
structure where it will live alongside a future DocumentManager
(management flow) and shared utilities.
