+++ src/components/Employee/OnboardingFlow/onboardingStateMachine.ts
@@ -13,7 +13,7 @@ import {
 import type { MachineEventType, MachineTransition } from '@/types/Helpers'
 import { CompensationContextual } from '@/components/Employee/Compensation'
 import { DeductionsContextual } from '@/components/Employee/OnboardingFlow/OnboardingFlowComponents'
-import { EmployeeDocumentsContextual } from '@/components/Employee/EmployeeDocuments'
+import { EmployeeDocumentsContextual } from '@/components/Employee/Documents/onboarding/EmployeeDocuments'
 import { PaymentMethodContextual } from '@/components/Employee/PaymentMethod'
 import { ProfileContextual } from '@/components/Employee/Profile/onboarding/Profile'
 import { OnboardingSummaryContextual } from '@/components/Employee/OnboardingSummary'
