+++ src/components/Employee/exports/employeeManagement.ts
@@ -1,5 +1,5 @@
 export { ManagementEmployeeList as EmployeeList } from '../EmployeeList/management/ManagementEmployeeList'
-export { EmployeeDocuments } from '../EmployeeDocuments'
+export { EmployeeDocuments } from '../Documents/onboarding/EmployeeDocuments'
 export { DashboardFlow } from '../Dashboard'
 export { WorkAddress } from '../WorkAddress/management/WorkAddress'
 export type { WorkAddressProps } from '../WorkAddress/management/WorkAddress'
