+++ src/components/Employee/index.ts
@@ -11,7 +11,7 @@ export { Landing } from './Landing'
 export { DocumentSigner } from './DocumentSigner'
 export { OnboardingFlow } from './OnboardingFlow/OnboardingFlow'
 export { SelfOnboardingFlow } from './SelfOnboardingFlow/SelfOnboardingFlow'
-export { EmployeeDocuments } from './EmployeeDocuments'
+export { EmployeeDocuments } from './Documents/onboarding/EmployeeDocuments'
 export { DashboardFlow } from './Dashboard'
 export type { DashboardFlowProps } from './Dashboard'
 export { HomeAddress } from './HomeAddress/management/HomeAddress'
